Make command generate all sitemaps and use progressbar for pretty output
